+++ modules/default/weatherforecast/node_helper.js
@@ -1,5 +1,5 @@
 const NodeHelper = require("node_helper");
-const Log = require("../../../js/logger");
+const Log = require("logger");
 
 module.exports = NodeHelper.create({
 	// Override start method.